Working

+++ RingSoft.HomeLogix.DataAccess/LookupModel/BankAccountLookup.cs
@@ -4,6 +4,6 @@
     {
         public string Description { get; set; }
 
-        public decimal CurrentBalance { get; set; }
+        public double CurrentBalance { get; set; }
     }
 }
