+++ RingSoft.HomeLogix.DataAccess/LookupModel/BudgetPeriodHistoryLookup.cs
@@ -8,10 +8,10 @@ namespace RingSoft.HomeLogix.DataAccess.LookupModel
 
         public DateTime PeriodEndingDate { get; set; }
 
-        public decimal ProjectedAmount { get; set; }
+        public double ProjectedAmount { get; set; }
 
-        public decimal ActualAmount { get; set; }
+        public double ActualAmount { get; set; }
 
-        public decimal Difference { get; set; }
+        public double Difference { get; set; }
     }
 }
