+++ RingSoft.HomeLogix.DataAccess/Model/History.cs
@@ -32,10 +32,10 @@ namespace RingSoft.HomeLogix.DataAccess.Model
         public string Description { get; set; }
 
         [Required]
-        public decimal ProjectedAmount { get; set; }
+        public double ProjectedAmount { get; set; }
 
         [Required]
-        public decimal ActualAmount { get; set; }
+        public double ActualAmount { get; set; }
 
         [MaxLength(250)]
         public string BankText { get; set; }
