+++ RingSoft.HomeLogix.MobileInterop/PhoneModel/BudgetStatistics.cs
@@ -13,16 +13,16 @@ namespace RingSoft.HomeLogix.Library.PhoneModel
 
         public DateTime MonthEnding { get; set; }
 
-        public decimal BudgetIncome { get; set; }
+        public double BudgetIncome { get; set; }
 
-        public decimal BudgetExpenses { get; set; }
+        public double BudgetExpenses { get; set; }
 
-        public decimal ActualIncome { get; set; }
+        public double ActualIncome { get; set; }
 
-        public decimal ActualExpenses { get; set; }
+        public double ActualExpenses { get; set; }
 
-        public decimal YtdIncome { get; set; }
+        public double YtdIncome { get; set; }
 
-        public decimal YtdExpenses { get; set; }
+        public double YtdExpenses { get; set; }
     }
 }
