+++ RingSoft.HomeLogix.MobileInterop/PhoneModel/HistoryData.cs
@@ -18,11 +18,11 @@ namespace RingSoft.HomeLogix.Library.PhoneModel
 
         public string Description { get; set; }
 
-        public decimal ProjectedAmount { get; set; }
+        public double ProjectedAmount { get; set; }
 
-        public decimal ActualAmount { get; set; }
+        public double ActualAmount { get; set; }
 
-        public decimal Difference { get; set; }
+        public double Difference { get; set; }
 
         public string BankText { get; set; }
